Code fixes for XLC and Clang execution.
Build system changes are not included for now.
